requirements.txt
asgiref==3.3.4
countries==0.2.0
Django==3.2.3
django-countries==7.2.1
django-multiselectfield==0.1.12
pytz==2021.1
sqlparse==0.4.1
